requirements.txt
ccxt
pandas
python-dotenv
ta-lib
numpy
aiohttp
streamlit
watchdog
tweepy
praw
textblob
binance-connector
requests
urllib3
